requirements.txt
numpy>=1.20.0
matplotlib>=3.4.0
scipy>=1.6.0
pandas>=1.2.0
geopandas>=0.9.0
rasterio>=1.2.0
pyyaml>=6.0
tqdm>=4.60.0
scikit-learn>=0.24.0
opencv-python>=4.5.0
gdal>=3.2.0
pyproj>=3.0.0
pyntcloud>=0.1.5
laspy>=2.0.0
plotly>=5.0.0
open3d>=0.13.0
pyCloudCompare>=0.1.5
shapely>=1.8.0
h5py>=3.6.0
jupyter>=1.0.0
seaborn>=0.11.0
scikit-image>=0.18.0 